add routes, add UI for comments, change controllers to work

+++ app/controllers/like_controller.rb
@@ -1,6 +1,7 @@
 class LikeController < ApplicationController
   def create
-    @like = Like.new(post_id: params[:like][:post_id], user: current_user)
+    @like = Post.find(params[:post_id]).likes.new(user: current_user)
+
     if @like.save
       redirect_to posts_path, notice: 'Post was liked'
     else
@@ -10,8 +11,8 @@ class LikeController < ApplicationController
 
   def destroy
     like = like.find(params[:id])
-    like.destroy
-    redirect_to posts_path, notice: 'unliked!'
+    like.destroy if like.user == current_user
+    redirect_to posts_path
   end
 
 end
